+++ neuron/src/lib/Neuron/Plugin/Type.hs
@@ -5,6 +5,7 @@
 
 module Neuron.Plugin.Type where
 
+import Data.Default
 import Data.Tagged
 import Neuron.Web.Route (NeuronWebT)
 import Neuron.Zettelkasten.Graph.Type (ZettelGraph)
@@ -12,6 +13,7 @@ import Neuron.Zettelkasten.ID (ZettelID)
 import Neuron.Zettelkasten.Resolver (ZIDRef)
 import Neuron.Zettelkasten.Zettel (ZettelC)
 import Reflex.Dom.Core (DomBuilder)
+import Reflex.Dom.Widget (blank)
 import Relude
 import qualified System.Directory.Contents.Types as DC
 
@@ -27,3 +29,12 @@ data Plugin pluginData = Plugin
     -- | Render this plugin's content below zettel html
     _plugin_renderPanel :: forall t m. DomBuilder t m => ZettelGraph -> pluginData -> NeuronWebT t m ()
   }
+
+instance Default (Plugin pluginData) where
+  def =
+    Plugin
+      { _plugin_filterSources = pure . Just,
+        _plugin_afterZettelRead = void . pure,
+        _plugin_afterZettelParse = id,
+        _plugin_renderPanel = const $ const blank
+      }