+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - Introducing *impulse* -- foundation for upcoming advanced search, replacing both z-index and legacy JS search (#108)
 - Remove (experimental) support for org-mode in v2 (use `v1` branch if you need org-mode)
+- Plugin support
+  - dirtree: Treat directories as zettels, forming automatic folgezettel connections to their children.
+  - neuronignore: Exclude directories and markdown files via `.neuronignore`.
 
 ## Unreleased (v1 + v2)